refactor(forcing): promote expand_yearly_files into jcm.forcing (#640)
Move the {year} file-pattern expansion out of runners into a public
jcm.forcing.expand_yearly_files (pure logic on plain types). runners keeps
the _expand_years alias so its many call sites are unchanged;
_product_available_years stays in runners as config reading. The pure
year-expansion tests move to forcing_test.py against the public name.
